+++ Examples/MAVMono/Online/src/location_manager.h
@@ -39,7 +39,7 @@ public:
     double interpolate(uint32_t x1, uint32_t x2, uint32_t x_predict, double y1, double y2);
     void stream_global_position(uint32_t timestamp, double lat, double lon, double alt);
 
-    uint64_t get_unixtime(uint32_t time);
+    uint64_t get_unixtime(uint64_t time);
     bool b_pixhawk_time_ref;
 
 private:
@@ -53,7 +53,7 @@ private:
     uint32_t c_local_timestamp, c_global_timestamp;
 
     // time reference from pixhawk
-    uint32_t pixhawk_ns_ref;
+    uint64_t pixhawk_ns_ref;
     uint64_t pixhawk_unix_ns_ref;
